+++ app/src/main/java/eu/kanade/tachiyomi/data/track/bangumi/BangumiApi.kt
@@ -45,6 +45,7 @@ class BangumiApi(
                 put("type", track.toApiStatus())
                 put("rate", track.score.toInt().coerceIn(0, 10))
                 put("ep_status", track.last_chapter_read.toInt())
+                put("private", track.private)
             }
                 .toString()
                 .toRequestBody()
@@ -62,6 +63,7 @@ class BangumiApi(
                 put("type", track.toApiStatus())
                 put("rate", track.score.toInt().coerceIn(0, 10))
                 put("ep_status", track.last_chapter_read.toInt())
+                put("private", track.private)
             }
                 .toString()
                 .toRequestBody()