+++ domain/src/main/java/tachiyomi/domain/track/model/Track.kt
@@ -16,4 +16,5 @@ data class Track(
     val remoteUrl: String,
     val startDate: Long,
     val finishDate: Long,
+    val private: Boolean,
 ) : Serializable